% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPICUREfiles.R
\name{EPICUREfiles}
\alias{EPICUREfiles}
\title{Liste les fichiers disponibles sur EPICURE avec un compte}
\usage{
EPICUREfiles(loginEpicure, passEpicure)
}
\arguments{
\item{loginEpicure}{votre login d'accès à EPICURE}

\item{passEpicure}{votre mot de passe pour EPICURE}
}
\value{
Un vecteur avec les noms des fichiers accessibles sur le FTP EPICURE avec le login et mot de passe.
}
\description{
Récupère les noms des fichiers disponibles sur EPICURE sur la base d'un login et d'un password
}
\examples{
EPICUREfiles("monlogin", "monpass")

}
