% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecupMeteo.R
\name{recupMeteo}
\alias{recupMeteo}
\title{Récupération des données météo}
\usage{
recupMeteo(loginEpicure, passEpicure, nomfichier)
}
\arguments{
\item{loginEpicure}{votre login d'accès à EPICURE}

\item{passEpicure}{votre mot de passe pour EPICURE}

\item{nomfichier}{le nom du fichier à consulter}
}
\value{
Un data frame avec les données météo pour chaque site.
}
\description{
Récupère les données météo sur la base d'un login, password et nom de fichier
}
\examples{
recupMeteo("monlogin", "monpass", "monfichier.csv")

}
